chore: Rename a common module
