+++ src/lib.rs
@@ -1,2 +1,2 @@
-pub mod common;
 pub mod messages;
+pub mod unix;
